Day 6 rewrite to BigInteger

+++ src/test/java/nl/maartenwiegers/aoc2021/Day6Test.java
@@ -2,6 +2,8 @@ package nl.maartenwiegers.aoc2021;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigInteger;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class Day6Test {
@@ -10,11 +12,11 @@ public class Day6Test {
 
     @Test
     public void shouldFindCorrectAnswerForDay6Part1() {
-        assertEquals(386536, day6.getCountOfFishAfterDays(80));
+        assertEquals(BigInteger.valueOf(386536), day6.getCountOfFishAfterDays(80));
     }
 
     @Test
     public void shouldFindCorrectAnswerForDay6Part2() {
-        assertEquals(1732821262171L, day6.getCountOfFishAfterDays(256));
+        assertEquals(BigInteger.valueOf(1732821262171L), day6.getCountOfFishAfterDays(256));
     }
 }
